requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.4
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
diffusers @ git+https://github.com/kashif/diffusers.git@143df09caf98ce00bdbd469294214acd8e577bc6
distro==1.9.0
docarray==0.40.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
faiss-cpu==1.7.4
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.19.1
gradio_client==0.10.0
grandalf==0.8
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
langchain==0.1.6
langchain-cli==0.0.21
langchain-community==0.0.21
langchain-core==0.1.25
langchain-openai==0.0.5
langchainhub==0.1.14
langcodes==3.3.0
langserve==0.0.41
langsmith==0.1.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.13
packaging==23.2
pandas==2.2.0
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
preshed==3.0.9
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
ruff==0.2.2
safetensors==0.4.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.26
srsly==2.4.8
sse-starlette==1.8.2
starlette==0.36.3
streamlit==1.31.1
sympy==1.12
tavily-python==0.3.1
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
textblob==0.17.1
thinc==8.2.3
tiktoken==0.5.2
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.23.2
validators==0.22.0
wasabi==1.1.2
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0
